Remove "Available to Lock" stats box and align stats with header
