+++ liquidity/lib/StatsBox/package.json
@@ -1,0 +1,13 @@
+{
+  "name": "@snx-v3/StatsBox",
+  "private": true,
+  "main": "index.ts",
+  "version": "0.0.1",
+  "dependencies": {
+    "@chakra-ui/icons": "^2.1.1",
+    "@chakra-ui/react": "^2.8.2",
+    "@snx-v3/SynthSkeleton": "workspace:*",
+    "@snx-v3/Tooltip": "workspace:*",
+    "react": "^18.2.0"
+  }
+}
