+++ liquidity/lib/StatsTotalLocked/package.json
@@ -1,0 +1,15 @@
+{
+  "name": "@snx-v3/StatsTotalLocked",
+  "private": true,
+  "main": "index.ts",
+  "version": "0.0.1",
+  "dependencies": {
+    "@chakra-ui/react": "^2.8.2",
+    "@snx-v3/Amount": "workspace:*",
+    "@snx-v3/StatsBox": "workspace:*",
+    "@snx-v3/useLiquidityPositions": "workspace:*",
+    "@snx-v3/useParams": "workspace:*",
+    "@synthetixio/wei": "^2.74.4",
+    "react": "^18.2.0"
+  }
+}
