+++ liquidity/lib/StatsTotalPnl/package.json
@@ -1,0 +1,19 @@
+{
+  "name": "@snx-v3/StatsTotalPnl",
+  "private": true,
+  "main": "index.ts",
+  "version": "0.0.1",
+  "dependencies": {
+    "@chakra-ui/react": "^2.8.2",
+    "@snx-v3/DebtAmount": "workspace:*",
+    "@snx-v3/StatsBox": "workspace:*",
+    "@snx-v3/useBlockchain": "workspace:*",
+    "@snx-v3/useCollateralPrices": "workspace:*",
+    "@snx-v3/useLiquidityPositions": "workspace:*",
+    "@snx-v3/useParams": "workspace:*",
+    "@snx-v3/usePythPrice": "workspace:*",
+    "@snx-v3/useRewards": "workspace:*",
+    "@synthetixio/wei": "^2.74.4",
+    "react": "^18.2.0"
+  }
+}
