+++ liquidity/lib/SynthSkeleton/package.json
@@ -1,0 +1,10 @@
+{
+  "name": "@snx-v3/SynthSkeleton",
+  "private": true,
+  "main": "index.ts",
+  "version": "0.0.1",
+  "dependencies": {
+    "@chakra-ui/react": "^2.8.2",
+    "react": "^18.2.0"
+  }
+}
