+++ liquidity/ui/package.json
@@ -25,6 +25,8 @@
     "@snx-v3/Multistep": "workspace:*",
     "@snx-v3/NumberInput": "workspace:*",
     "@snx-v3/RepayModal": "workspace:*",
+    "@snx-v3/StatsTotalLocked": "workspace:*",
+    "@snx-v3/StatsTotalPnl": "workspace:*",
     "@snx-v3/TermsModal": "workspace:*",
     "@snx-v3/TokenIcon": "workspace:*",
     "@snx-v3/Tooltip": "workspace:*",
@@ -50,7 +52,6 @@
     "@snx-v3/useClearDebt": "workspace:*",
     "@snx-v3/useClosePosition": "workspace:*",
     "@snx-v3/useCollateralPriceUpdates": "workspace:*",
-    "@snx-v3/useCollateralPrices": "workspace:*",
     "@snx-v3/useCollateralTypes": "workspace:*",
     "@snx-v3/useContractErrorParser": "workspace:*",
     "@snx-v3/useCoreProxy": "workspace:*",
@@ -72,7 +73,6 @@
     "@snx-v3/usePoolsList": "workspace:*",
     "@snx-v3/usePositionManagerForCollateral": "workspace:*",
     "@snx-v3/usePythFeeds": "workspace:*",
-    "@snx-v3/usePythPrice": "workspace:*",
     "@snx-v3/usePythVerifier": "workspace:*",
     "@snx-v3/useRates": "workspace:*",
     "@snx-v3/useRepay": "workspace:*",
